EoDW notes bellow:
Need to create a function that hides check button and puts next button in so that country can be randomly generated once again. Also need to find a way to randomly generate one row from csv file and seperate that into seperate lists so that I can display the country, photo and capital seperatly as question and answer. Code right now has ranom generation error in function so that code can be run without breaking, however the code that is run has no effect at this time.