Notes bellow
In list each 'chosen' is a seperate variable (not one the can be seperated) meaning they are not linked and are just added to the list without being under the same number.
So the linked country, capital and image are all "chosen" but are definied as different variables within the list.
NOTE TO SELF:
REMOVE CSV FIRST LINE FROM FILE IN CODE